add forgotPassword and resetPassword
