+++ src/components/app/app.tsx
@@ -19,7 +19,7 @@ import { useEffect } from 'react';
 import { getIngredients } from '../../slices/ingredientsSlice';
 import { getFeed } from '../../slices/feedSlice';
 import { ProtectedRoute } from '../protected-routes/protected-route';
-import { checkUserAuth } from '../../slices/userSlice';
+import { checkUserAuth } from '../../slices/userSlice/authThunks';
 
 const App = () => {
   const dispatch = useDispatch();
