+++ src/components/profile-menu/profile-menu.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { useLocation } from 'react-router-dom';
 import { Preloader, ProfileMenuUI } from '@ui';
 import { useDispatch, useSelector } from '../../services/store';
-import { logoutUser } from '../../slices/userSlice';
+import { logoutUser } from '../../slices/userSlice/authThunks';
 
 export const ProfileMenu: FC = () => {
   const { pathname } = useLocation();
