+++ src/pages/forgot-password/forgot-password.tsx
@@ -1,30 +1,28 @@
 import { FC, useState, SyntheticEvent } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { forgotPasswordApi } from '@api';
 import { ForgotPasswordUI } from '@ui-pages';
+import { useDispatch, useSelector } from '../../services/store';
+import { forgotPassword } from '../../slices/userSlice/authThunks';
 
 export const ForgotPassword: FC = () => {
   const [email, setEmail] = useState('');
-  const [error, setError] = useState<Error | null>(null);
+  const dispatch = useDispatch();
+  const { error } = useSelector((state) => state.user);
 
   const navigate = useNavigate();
 
   const handleSubmit = (e: SyntheticEvent) => {
     e.preventDefault();
 
-    setError(null);
-    forgotPasswordApi({ email })
-      .then(() => {
-        localStorage.setItem('resetPassword', 'true');
-        navigate('/reset-password', { replace: true });
-      })
-      .catch((err) => setError(err));
+    dispatch(forgotPassword({ email })).then(() => {
+      navigate('/reset-password', { replace: true });
+    });
   };
 
   return (
     <ForgotPasswordUI
-      errorText={error?.message}
+      errorText={error}
       email={email}
       setEmail={setEmail}
       handleSubmit={handleSubmit}
