+++ src/pages/login/login.tsx
@@ -1,7 +1,7 @@
 import { FC, SyntheticEvent, useState } from 'react';
 import { LoginUI } from '@ui-pages';
 import { useDispatch, useSelector } from '../../services/store';
-import { loginUser } from '../../slices/userSlice';
+import { loginUser } from '../../slices/userSlice/authThunks';
 import { useNavigate } from 'react-router-dom';
 import { Preloader } from '@ui';
 
@@ -10,7 +10,7 @@ export const Login: FC = () => {
   const [password, setPassword] = useState('');
   const dispatch = useDispatch();
   const navigate = useNavigate();
-  const { error, loading } = useSelector((state) => state.user);
+  const { error } = useSelector((state) => state.user);
 
   const handleSubmit = (e: SyntheticEvent) => {
     e.preventDefault();
@@ -27,10 +27,6 @@ export const Login: FC = () => {
       }
     });
   };
-
-  if (loading) {
-    return <Preloader />;
-  }
 
   return (
     <LoginUI
