+++ src/slices/userSlice/userSlice.ts
@@ -1,81 +1,28 @@
+import { PayloadAction, createSlice } from '@reduxjs/toolkit';
+import { TOrder, TUser } from '@utils-types';
 import {
-  TLoginData,
-  TRegisterData,
-  getUserApi,
-  loginUserApi,
-  logoutApi,
-  registerUserApi,
-  updateUserApi
-} from '@api';
-import { PayloadAction, createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { TUser } from '@utils-types';
-import { deleteCookie, getCookie, setCookie } from '../utils/cookie';
-
-export const loginUser = createAsyncThunk(
-  'user/loginUser',
-  async ({ email, password }: TLoginData) => {
-    const res = await loginUserApi({ email, password });
-    setCookie('accessToken', res.accessToken);
-    localStorage.setItem('refreshToken', res.refreshToken);
-    return res;
-  }
-);
-export const logoutUser = createAsyncThunk('user/logoutUser', async () => {
-  const res = await logoutApi();
-  deleteCookie('accessToken');
-  localStorage.clear();
-  return res;
-});
-
-export const registerUser = createAsyncThunk(
-  'user/registerUser',
-  async ({ email, name, password }: TRegisterData) => {
-    const res = await registerUserApi({ email, name, password });
-    setCookie('accessToken', res.accessToken);
-    localStorage.setItem('refreshToken', res.refreshToken);
-    return res;
-  }
-);
-
-export const updateUser = createAsyncThunk(
-  'user/updateUser',
-  async ({ email, name, password }: Partial<TRegisterData>) => {
-    await updateUserApi({ email, name, password });
-    return getUserApi();
-  }
-);
-
-export const checkUserAuth = createAsyncThunk(
-  'user/checkUser',
-  async (_, { dispatch }) => {
-    if (getCookie('accessToken')) {
-      getUserApi()
-        .then((res) => setUser(res.user))
-        .catch(() => {
-          localStorage.removeItem('refreshToken');
-          deleteCookie('accessToken');
-        })
-        .finally(() => {
-          dispatch(authChecked());
-        });
-    } else {
-      dispatch(authChecked());
-    }
-  }
-);
+  loginUser,
+  registerUser,
+  updateUser,
+  logoutUser,
+  forgotPassword,
+  resetPassword
+} from './authThunks';
 
 type TUserState = {
   isAuthChecked: boolean;
   loading: boolean;
   user: TUser | null;
   error?: string | null;
+  orders?: TOrder[];
 };
 
 const initialState: TUserState = {
   isAuthChecked: false,
   loading: false,
   user: null,
-  error: null
+  error: null,
+  orders: []
 };
 
 const userSlice = createSlice({
@@ -148,6 +95,31 @@ const userSlice = createSlice({
       .addCase(logoutUser.fulfilled, (state) => {
         state.loading = false;
         state.user = null;
+      });
+    builder
+      .addCase(forgotPassword.pending, (state) => {
+        state.loading = true;
+        state.error = null;
+      })
+      .addCase(forgotPassword.rejected, (state, action) => {
+        state.loading = false;
+        state.error = action.error.message;
+      })
+      .addCase(forgotPassword.fulfilled, (state) => {
+        state.loading = false;
+      });
+
+    builder
+      .addCase(resetPassword.pending, (state) => {
+        state.loading = true;
+        state.error = null;
+      })
+      .addCase(resetPassword.rejected, (state, action) => {
+        state.loading = false;
+        state.error = action.error.message;
+      })
+      .addCase(resetPassword.fulfilled, (state) => {
+        state.loading = false;
       });
   }
 });
